Automation

+++ trigger_swann_from_hrdsummary.sh
@@ -7,6 +7,7 @@ outputString=$(python hot_run_from_hrdsummary.py --path $1)
 mapfile -t output_lines <<< "$outputString"
 
 # set var names
+dataexist=${output_lines[-8]}
 filesexist=${output_lines[-7]}
 tc=${output_lines[-6]}
 stormid=${output_lines[-5]}
@@ -17,9 +18,11 @@ lon=${output_lines[-1]} # not using right now, might add later ******
 
 # also add in center location? ******
 # add check for winter storms!! ******
-if [[ "$filesexist" == 'N' && "$tc" == 'TC' ]]; then 
+if [[ "$filesexist" == 'N' && "$tc" == 'TC' && "$dataexist" == 'True' ]]; then 
     echo "python hot_main_run_samurai.py $stormid $legstart $legend > ./output_files/sam_$stormid_$legstart.log"
     python hot_main_run_samurai.py $stormid $legstart $legend > ./output_files/sam_$stormid_$legstart.log
+elif [[ "$filesexist" == 'N' && "$tc" == 'TC' && "$dataexist" == 'False' ]]; then 
+    echo "files don't exist, but waiting for more data."
 else
     echo "files already processed: $stormid $legstart $legend"
 fi
